requirements.txt
Flask==3.1.2
google-generativeai==0.8.5
google-auth-oauthlib==1.2.1
google-genai==1.46.0
SQLAlchemy==2.0
python-dotenv==1.2.1